// lib/interfaces/si.interface.ts
export interface SIDistance {
    /**
     * @description Distance in millimeters
     */
    millimeters: number;
    /**
     * @description Distance in centimeters
     */
    centimeters: number;
    /**
     * @description Distance in meters
     */
    meters: number;
}

export interface SITime {
    /**
     * @description Time in nanoseconds
     */
    nanoseconds: number;
    /**
     * @description Time in microseconds
     */
    microseconds: number;
    /**
     * @description Time in milliseconds
     */
    milliseconds: number;
    /**
     * @description Time in seconds
     */
    seconds: number;
    /**
     * @description Time in minutes
     */
    minutes: number;
    /**
     * @description Time in hours
     */
    hours: number;
}

export interface SIVelocity {
    /**
     * @description Velocity in milimeters per nanosecond
     */
    milimeters_per_nanosecond: number;
    /**
     * @description Velocity in centimeters per nanosecond
     */
    centimeters_per_nanosecond: number;
    /**
     * @description Velocity in meters per nanosecond
     */
    meters_per_nanosecond: number;
    /**
     * @description Velocity in milimeters per microsecond
     */
    milimeters_per_microsecond: number;
    /**
     * @description Velocity in centimeters per microsecond
     */
    centimeters_per_microsecond: number;
    /**
     * @description Velocity in meters per microsecond
     */
    meters_per_microsecond: number;
    /**
     * @description Velocity in milimeters per millisecond
     */
    milimeters_per_millisecond: number;
    /**
     * @description Velocity in centimeters per millisecond
     */
    centimeters_per_millisecond: number;
    /**
     * @description Velocity in meters per millisecond
     */
    meters_per_millisecond: number;
    /**
     * @description Velocity in milimeters per second
     */
    milimeters_per_second: number;
    /**
     * @description Velocity in centimeters per second
     */
    centimeters_per_second: number;
    /**
     * @description Velocity in meters per second
     */
    meters_per_second: number;
    /**
     * @description Velocity in milimeters per minute
     */
    milimeters_per_minute: number;
    /**
     * @description Velocity in centimeters per minute
     */
    centimeters_per_minute: number;
    /**
     * @description Velocity in meters per minute
     */
    meters_per_minute: number;
    /**
     * @description Velocity in milimeters per hour
     */
    milimeters_per_hour: number;
    /**
     * @description Velocity in centimeters per hour
     */
    centimeters_per_hour: number;
    /**
     * @description Velocity in meters per hour
     */
    meters_per_hour: number;
}