// lib/interfaces/velocity.interface.ts
export type VelocityUnit =
    | 'milimeters_per_nanosecond'
    | 'centimeters_per_nanosecond'
    | 'meters_per_nanosecond'
    | 'milimeters_per_microsecond'
    | 'centimeters_per_microsecond'
    | 'meters_per_microsecond'
    | 'milimeters_per_millisecond'
    | 'centimeters_per_millisecond'
    | 'meters_per_millisecond'
    | 'milimeters_per_second'
    | 'centimeters_per_second'
    | 'meters_per_second'
    | 'milimeters_per_minute'
    | 'centimeters_per_minute'
    | 'meters_per_minute'
    | 'milimeters_per_hour'
    | 'centimeters_per_hour'
    | 'meters_per_hour';